backend/apps/m3_webrtc/pyproject.toml
[project]
name = "HyperSight_WebRTC"
version = "0.0.0"
authors = [{ name = "tungp99", email = "tung.cloud@outlook.com" }]
requires-python = ">=3.13,<3.14"
dependencies = [
  "aiohttp>=3.11.14",
  "aiortc>=1.11.0",
  "ffmpeg-python>=0.2.0",
  "HyperSight_Configuration",
  "numpy>=2.2.4",
  "pygobject>=3.52.3",
  "websockets>=15.0.1",
]

[tool.uv.sources]
HyperSight_Configuration = { workspace = true }